requirements.txt
tensorboard
matplotlib
seaborn
torchmetrics
pytorch-lightning
fairscale
datasets
einops
#pip install torch==1.12.1+cu113 torchvision==0.13.1+cu113 torchaudio==0.12.1 --extra-index-url https://download.pytorch.org/whl/cu113